requirements.txt
asgi-redis==0.14.1
asgiref==0.14.0
autobahn==0.16.0
channels==0.17.3
colorama==0.3.7
contextlib2==0.5.4
coverage==4.2
daphne==0.14.3
dj-database-url==0.4.1
Django==1.8.8
django-countries==4.0
django-floppyforms==1.7.0
django-idmap==1.0
django-vanilla-views==1.0.4
djangorestframework==3.4.6
gunicorn==19.7.1
honcho==0.7.1
huey==1.2.0
IPy==0.83
msgpack-python==0.4.8
otree-boto2-shim==0.3.2
otree-core==1.4.31
otree-save-the-change==1.1.3
pbr==1.10.0
psycopg2>=2.5.1
py==1.4.31
pytest==2.9.2
pytest-django==3.0.0
python-redis-lock==3.2.0
pytz==2016.6.1
raven==5.25.0
redis==2.10.5
requests==2.11.1
schema==0.6.2
six==1.10.0
Twisted==16.2.0
txaio==2.5.1
unicodecsv==0.14.1
whitenoise==3.2.1
ws4py==0.3.5
XlsxWriter==0.9.3
zope.interface==4.2.0
